Here's a summary of the recent code changes I made:
refactor: Use `@web.worker/s3-upload-folder` for MP3 uploads

This commit refactors the MP3-to-S3 workflow to simplify the process by using the `@web.worker/s3-upload-folder` CLI tool.

Key changes:

1.  **MP3 Download Script (`scripts/downloadMp3sToLocal.ts`):**
    -   Formerly `scripts/downloadMp3sToS3.ts`, this script now solely downloads MP3 files from URLs in the weekly JSON to a local `./media/` directory.
    -   All direct S3 upload logic and AWS SDK dependencies have been removed from this script.

2.  **GitHub Action Workflow (`.github/workflows/download-mp3s.yml`):**
    -   The workflow now orchestrates a two-step process:
        1.  Creates a `./media/` directory.
        2.  Runs `scripts/downloadMp3sToLocal.ts` to populate `./media/` with downloaded MP3s.
        3.  Executes `npx @web.worker/s3-upload-folder` to upload the contents of the `./media/` directory to the S3-compatible storage.
    -   S3 configuration (endpoint, credentials, bucket, region, prefix) is passed as arguments to the CLI tool from GitHub secrets.

3.  **Dependencies (`package.json`):**
    -   Removed the `@aws-sdk/client-s3` dependency as it's no longer needed.
    -   The `download-mp3s` script in `package.json` now points to `scripts/downloadMp3sToLocal.ts`.

4.  **Documentation (`README.md`):**
    -   Updated to reflect the new workflow, detailing the use of `scripts/downloadMp3sToLocal.ts` for local downloads and `@web.worker/s3-upload-folder` for the S3 upload.
    -   Clarified the S3 secrets required for the CLI tool.

This change simplifies the custom scripting involved in S3 uploads by leveraging an external specialized CLI tool.
