--- .github/workflows/update-weekly-json.yml
@@ -1,37 +1,0 @@
-name: Update Weekly Podcast JSON
-
-on:
-  schedule:
-    - cron: '0 0 * * 1' # Run at 00:00 UTC every Monday
-  workflow_dispatch: # Allows manual triggering
-
-jobs:
-  update_json:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write # Needed to commit changes
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Bun
-        uses: oven-sh/setup-bun@v1
-        with:
-          bun-version: latest # or specify a version like "1.0.0"
-
-      - name: Install dependencies
-        run: bun install --frozen-lockfile
-
-      - name: Run update script
-        run: bun run ./scripts/updateWeeklyJson.ts # Adjusted path for clarity
-
-      - name: Commit and push changes
-        uses: stefanzweifel/git-auto-commit-action@v5
-        with:
-          commit_message: 'chore: Update weekly podcast JSON'
-          branch: main # Or your default branch
-          file_pattern: 'results/*.json' # Only commit changes in the results directory
-          commit_user_name: 'github-actions[bot]'
-          commit_user_email: 'github-actions[bot]@users.noreply.github.com'
-          commit_author: 'github-actions[bot] <github-actions[bot]@users.noreply.github.com>'
-```
